# packages/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact: {}
  description: Snapflow AI platform API Docs
  title: Snapflow
  version: "1.0"
servers:
- url: http://localhost:3000
paths:
...
  /organizations/invitations:
    get:
      operationId: listOrganizationInvitationsForAuthenticatedUser
      parameters: []
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/OrganizationInvitation'
                type: array
          description: List of organization invitations
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: List organization invitations for authenticated user
      tags:
      - organizations
  /organizations/invitations/count:
    get:
      operationId: getOrganizationInvitationsCountForAuthenticatedUser
      parameters: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: number
          description: Count of organization invitations
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get count of organization invitations for authenticated user
      tags:
      - organizations
  /organizations/invitations/{invitationId}/accept:
    post:
      operationId: acceptOrganizationInvitation
      parameters:
      - description: Invitation ID
        explode: false
        in: path
        name: invitationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Organization invitation accepted successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Accept organization invitation
      tags:
      - organizations
  /organizations/invitations/{invitationId}/decline:
    post:
      operationId: declineOrganizationInvitation
      parameters:
      - description: Invitation ID
        explode: false
        in: path
        name: invitationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Organization invitation declined successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Decline organization invitation
      tags:
      - organizations
  /organizations:
    get:
      operationId: listOrganizations
      parameters: []
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Organization'
                type: array
          description: List of organizations
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: List organizations
      tags:
      - organizations
    post:
      operationId: createOrganization
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganization'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: Organization created successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Create organization
      tags:
      - organizations
  /organizations/{organizationId}:
    delete:
      operationId: deleteOrganization
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Organization deleted successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Delete organization
      tags:
      - organizations
    get:
      operationId: getOrganization
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: Organization details
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get organization by ID
      tags:
      - organizations
  /organizations/{organizationId}/usage:
    get:
      operationId: getOrganizationUsageOverview
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageOverview'
          description: Current usage overview
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get organization current usage overview
      tags:
      - organizations
  /organizations/{organizationId}/quota:
    patch:
      operationId: updateOrganizationQuota
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationQuota'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: Organization details
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Update organization quota
      tags:
      - organizations
  /organizations/{organizationId}/leave:
    post:
      operationId: leaveOrganization
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Organization left successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Leave organization
      tags:
      - organizations
  /organizations/{organizationId}/suspend:
    post:
      operationId: suspendOrganization
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationSuspension'
        required: false
      responses:
        "204":
          description: Organization suspended successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Suspend organization
      tags:
      - organizations
  /organizations/{organizationId}/unsuspend:
    post:
      operationId: unsuspendOrganization
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Organization unsuspended successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Unsuspend organization
      tags:
      - organizations
...
  /organizations/{organizationId}/invitations:
    get:
      operationId: listOrganizationInvitations
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/OrganizationInvitation'
                type: array
          description: List of pending organization invitations
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: List pending organization invitations
      tags:
      - organizations
    post:
      operationId: createOrganizationInvitation
      parameters:
      - description: Organization ID
        explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationInvitation'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationInvitation'
          description: Organization invitation created successfully
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Create organization invitation
      tags:
      - organizations
...
components:
  schemas:
...
    OrganizationRole:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        permissions:
        - write:registries
        - write:registries
        name: name
        isGlobal: true
        description: description
        id: id
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Role ID
          type: string
        name:
          description: Role name
          type: string
        description:
          description: Role description
          type: string
        permissions:
          description: Roles assigned to the user
          items:
            enum:
            - write:registries
            - delete:registries
            - write:images
            - delete:images
            - write:sandboxes
            - delete:sandboxes
            - read:buckets
            - write:buckets
            - delete:buckets
            type: string
          type: array
        isGlobal:
          description: Global role flag
          type: boolean
        createdAt:
          description: Creation timestamp
          format: date-time
          type: string
        updatedAt:
          description: Last update timestamp
          format: date-time
          type: string
      required:
      - createdAt
      - description
      - id
      - isGlobal
      - name
      - permissions
      - updatedAt
      type: object
    OrganizationInvitation:
      example:
        organizationId: organizationId
        createdAt: 2000-01-23T04:56:07.000+00:00
        invitedBy: invitedBy
        role: owner
        organizationName: organizationName
        assignedRoles:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          permissions:
          - write:registries
          - write:registries
          name: name
          isGlobal: true
          description: description
          id: id
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          permissions:
          - write:registries
          - write:registries
          name: name
          isGlobal: true
          description: description
          id: id
          updatedAt: 2000-01-23T04:56:07.000+00:00
        id: id
        email: email
        expiresAt: 2000-01-23T04:56:07.000+00:00
        status: pending
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Invitation ID
          type: string
        email:
          description: Email address of the user being invited
          type: string
        invitedBy:
          description: Email address of the inviter
          type: string
        organizationId:
          description: Organization ID
          type: string
        organizationName:
          description: Organization name
          type: string
        expiresAt:
          description: Expiration date of the invitation
          format: date-time
          type: string
        status:
          description: Invitation status
          enum:
          - pending
          - accepted
          - declined
          - cancelled
          type: string
        role:
          description: Member role
          enum:
          - owner
          - member
          type: string
        assignedRoles:
          description: Assigned roles
          items:
            $ref: '#/components/schemas/OrganizationRole'
          type: array
        createdAt:
          description: Creation timestamp
          format: date-time
          type: string
        updatedAt:
          description: Last update timestamp
          format: date-time
          type: string
      required:
      - assignedRoles
      - createdAt
      - email
      - expiresAt
      - id
      - invitedBy
      - organizationId
      - organizationName
      - role
      - status
      - updatedAt
      type: object
    CreateOrganization:
      example:
        name: Example organization
      properties:
        name:
          description: Name of the organization
          example: Example organization
          type: string
      required:
      - name
      type: object
    Organization:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        suspendedUntil: 2000-01-23T04:56:07.000+00:00
        createdBy: createdBy
        suspensionReason: suspensionReason
        suspendedAt: 2000-01-23T04:56:07.000+00:00
        name: name
        personal: true
        id: id
        suspended: true
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Organization ID
          type: string
        name:
          description: Organization name
          type: string
        createdBy:
          description: The id of the organization creator
          type: string
        personal:
          description: Personal organization flag
          type: boolean
        createdAt:
          description: Creation timestamp
          format: date-time
          type: string
        updatedAt:
          description: Last update timestamp
          format: date-time
          type: string
        suspended:
          description: Suspended flag
          type: boolean
        suspendedAt:
          description: Suspended at
          format: date-time
          type: string
        suspensionReason:
          description: Suspended reason
          type: string
        suspendedUntil:
          description: Suspended until
          format: date-time
          type: string
      required:
      - createdAt
      - createdBy
      - id
      - name
      - personal
      - suspended
      - suspendedAt
      - suspendedUntil
      - suspensionReason
      - updatedAt
      type: object
    UsageOverview:
      example:
        totalCpuQuota: 0.8008281904610115
        currentDiskUsage: 7.061401241503109
        currentMemoryUsage: 2.3021358869347655
        currentCpuUsage: 5.637376656633329
        totalGpuQuota: 6.027456183070403
        totalMemoryQuota: 1.4658129805029452
        totalDiskQuota: 5.962133916683182
      properties:
        totalCpuQuota:
          type: number
        totalGpuQuota:
          type: number
        totalMemoryQuota:
          type: number
        totalDiskQuota:
          type: number
        currentCpuUsage:
          type: number
        currentMemoryUsage:
          type: number
        currentDiskUsage:
          type: number
      required:
      - currentCpuUsage
      - currentDiskUsage
      - currentMemoryUsage
      - totalCpuQuota
      - totalDiskQuota
      - totalGpuQuota
      - totalMemoryQuota
      type: object
    UpdateOrganizationQuota:
      example:
        totalCpuQuota: 0.8008281904610115
        maxCpuPerSandbox: 5.962133916683182
        imageQuota: 7.061401241503109
        maxDiskPerSandbox: 2.3021358869347655
        maxMemoryPerSandbox: 5.637376656633329
        maxImageSize: 9.301444243932576
        totalMemoryQuota: 6.027456183070403
        bucketQuota: 3.616076749251911
        totalDiskQuota: 1.4658129805029452
      properties:
        totalCpuQuota:
          nullable: true
          type: number
        totalMemoryQuota:
          nullable: true
          type: number
        totalDiskQuota:
          nullable: true
          type: number
        maxCpuPerSandbox:
          nullable: true
          type: number
        maxMemoryPerSandbox:
          nullable: true
          type: number
        maxDiskPerSandbox:
          nullable: true
          type: number
        imageQuota:
          nullable: true
          type: number
        maxImageSize:
          nullable: true
          type: number
        bucketQuota:
          nullable: true
          type: number
      required:
      - bucketQuota
      - imageQuota
      - maxCpuPerSandbox
      - maxDiskPerSandbox
      - maxImageSize
      - maxMemoryPerSandbox
      - totalCpuQuota
      - totalDiskQuota
      - totalMemoryQuota
      type: object
    OrganizationSuspension:
      example:
        reason: reason
        until: 2000-01-23T04:56:07.000+00:00
      properties:
        reason:
          description: Suspension reason
          type: string
        until:
          description: Suspension until
          format: date-time
          type: string
      required:
      - reason
      - until
      type: object
...
    CreateOrganizationInvitation:
      example:
        role: member
        assignedRoleIds:
        - assignedRoleIds
        - assignedRoleIds
        email: test@mail.com
        expiresAt: 2025-07-03T23:59:59Z
      properties:
        email:
          description: Email address of the inviting user
          example: test@mail.com
          type: string
        role:
          default: member
          description: Organization role of the inviting user
          enum:
          - owner
          - member
          type: string
        assignedRoleIds:
          default:
          - 00000000-0000-0000-0000-000000000001
          description: Role IDs of the inviting user
          items:
            type: string
          type: array
        expiresAt:
          description: Expiration date of the invitation
          example: 2025-07-03T23:59:59Z
          format: date-time
          type: string
      required:
      - assignedRoleIds
      - email
      - role
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http